Reading and writing with Jaxb now performed * Needs arcpoints adding * Needs a few bugs fixing such as refreshing tokens

+++ pipe-core/src/main/java/parser/ExprEvaluator.java
@@ -116,7 +116,7 @@ public class ExprEvaluator {
         String capacityWithPlaceName =
                 "cap(" + place.getName().replaceAll("\\s", "") + ")";
         if (expr.toLowerCase().contains(capacityWithPlaceName.toLowerCase())) {
-            double capacity = place.getCapacity();
+            int capacity = place.getCapacity();
             expr = expr.toLowerCase()
                     .replace(capacityWithPlaceName.toLowerCase(),
                             Double.toString(capacity));
